src/main/java/algorithms/leetcode/HeightChecker.java
package algorithms.leetcode;

import java.util.Arrays;

/**
 * @link <a href="https://leetcode.com/problems/height-checker/">https://leetcode.com/problems/height-checker/</a>
 * @date 15.05.2023
 */
public class HeightChecker {
    public int heightChecker(int[] heights) {
        int[] copy = Arrays.copyOf(heights, heights.length);
        Arrays.sort(copy);
        int count = 0;
        for (int i = 0; i < heights.length; i++) {
            if (heights[i] != copy[i]) {
                count++;
            }
        }
        return count;
    }
}
